Avoid trying to connect to master twice if unavailable
The method #outside_transaction? gets called during the exception
handling of transaction errors. In case a master connection is not
available and therefore, a transaction begin failed, it's not desired to
delegate the #outside_transaction? method to the master again as it
would trigger another connect.

The postgres adapter is currently the only adapter which supports this
method. A possible implementation of the MasterSlave approach for
PostgreSQL would need to overrite this method again.